pyproject.toml
[project]
name = "warpkit"
description = "A python library for neuroimaging transformations"
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT License" }
authors = [{ name = "Andrew Van", email = "vanandrew@wustl.edu" }]
keywords = ["neuroimaging"]
classifiers = ["License :: OSI Approved :: MIT License", "Programming Language :: Python :: 3"]
urls = { github = "https://github.com/vanandrew/warpkit" }
version = "0.0.1"
dynamic = ["entry-points"]
dependencies = [
  "nibabel >= 4.0.2",
  "numpy >= 1.23.3",
  "scikit-image >= 0.19.3",
  "scipy >= 1.8.1",
  "pybids >= 0.15.1",
  "memori >= 0.2.1",
  "transforms3d >= 0.4.1",
  "indexed-gzip >= 1.7.0",
]

[project.optional-dependencies]
dev = ["black >= 22.3.0", "coverage[toml] >= 5.5", "pytest >= 6.2.5", "pytest-pycodestyle >= 2.2.0"]

[build-system]
requires = ["setuptools", "wheel", "cmake >= 3.13", "cmake-setuptools-ext >= 0.1.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
zip-safe = true

[tool.setuptools.packages.find]
exclude = ["tests"]

[tool.black]
line-length = 120
target-version = ["py37", "py38", "py39", "py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
